feat: user api route to get saved content for the user
