+++ server/api/users.js
@@ -1,17 +1,26 @@
 const router = require('express').Router()
-const {User} = require('../db/models')
+const {User, Content, Interest} = require('../db/models')
 module.exports = router
 
-router.get('/', async (req, res, next) => {
+router.get('/:userId/content', async (req, res, next) => {
   try {
-    const users = await User.findAll({
-      // explicitly select only the id and email fields - even though
-      // users' passwords are encrypted, it won't help if we just
-      // send everything to anyone who asks!
-      attributes: ['id', 'firstName', 'lastName', 'username', 'email']
+    let userId = Number(req.params.userId)
+    let savedContent = await User.findAll({
+      where: {
+        id: userId
+      },
+      attributes: ['id'],
+      include: [
+        {
+          model: Content,
+          attributes: ['title', 'sourceUrl', 'imageUrl', 'description'],
+          include: [{model: Interest, attributes: ['id', 'name']}]
+        }
+      ]
     })
-    res.json(users)
+
+    res.status(200).send(savedContent)
   } catch (err) {
-    next(err)
+    console.error(err)
   }
 })
